+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# unreleased
+
+- Change download link label of attachments from "Download" to the actual filename
+- Support multiple attachments by providing a simple UI for adding and removing attachments
+
 # 0.6.1
 
 - Implement `default_sorting` in `Compony::Components::List`